+++ Docker/cron/docassemble-cron-daily.sh
@@ -124,8 +124,7 @@ if [ "${DABACKUPDAYS}" != "0" ]; then
 	    cp /var/lib/redis/dump.rdb "${BACKUPDIR}/redis.rdb"
         fi
     elif [[ $CONTAINERROLE =~ .*:cron:.* ]] && [ "${REDIS:-redis://localhost}" != "redis://localhost" ]; then
-	REDISHOST="$(echo ${REDIS} | cut -c9-)"
-	redis-cli -h "${REDISHOST}" --rdb "${BACKUPDIR}/redis.rdb"
+	$REDISCLI --rdb "${BACKUPDIR}/redis.rdb"
     fi
 
     if [[ $CONTAINERROLE =~ .*:(all|sql):.* ]]; then