+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.5.67] - 2019-12-19
+### Added
+- The `help generator` and `image generator` field modifiers.
+### Fixed
+- Error when `code` was used with `content file` and the code 
+  returned a list.
+- Backup of Redis database when password is used.
+
 ## [0.5.66] - 2019-12-19
 ### Fixed
 - Issue with Redis and passwords.